+++ src/main/java/com/mesilat/vbp/impl/ValidatorResource.java
@@ -1,10 +1,16 @@
 package com.mesilat.vbp.impl;
 
-import com.mesilat.vbp.api.Validator;
+import com.atlassian.confluence.security.PermissionManager;
 import com.atlassian.plugin.spring.scanner.annotation.component.Scanned;
+import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
+import com.atlassian.sal.api.message.I18nResolver;
+import com.atlassian.sal.api.websudo.WebSudoNotRequired;
+import com.atlassian.sal.api.websudo.WebSudoRequired;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.mesilat.vbp.Constants;
+import com.mesilat.vbp.api.Validator;
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -25,17 +31,20 @@ import java.util.List;
 
 @Path("/validator")
 @Scanned
-public class ValidatorResource {
-    private static final Logger LOGGER = LoggerFactory.getLogger("com.mesilat.vbp");
+@WebSudoRequired
+public class ValidatorResource extends ResourceBase {
+    private static final Logger LOGGER = LoggerFactory.getLogger(Constants.PLUGIN_KEY);
 
+    private final I18nResolver resolver;
     private final ValidatorManager service;
     private final ObjectMapper mapper = new ObjectMapper();
 
     @GET
     @Path("/")
     @Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
+    @WebSudoNotRequired
     public Response get(@QueryParam("extensive") Boolean extensive) {
-        LOGGER.debug("List validators");
+        LOGGER.trace("List validators");
         try {
             List<Validator> validators = service.list(extensive == null? false: extensive);
             String serialized = mapper.writerWithDefaultPrettyPrinter().writeValueAsString(validators);
@@ -48,16 +57,18 @@ public class ValidatorResource {
     @GET
     @Path("/css")
     @Produces("text/css;charset=utf-8")
+    @WebSudoNotRequired
     public Response css() {
-        LOGGER.debug("List validators");
+        LOGGER.trace("Get validators CSS");
         return Response.ok(service.css()).build();
     }
     
     @GET
     @Path("/{code}")
     @Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
+    @WebSudoNotRequired
     public Response get(@PathParam("code") String code) {
-        LOGGER.debug(String.format("Get validator info for %s", code));
+        LOGGER.trace(String.format("Get validator %s", code));
 
         Validator validator = service.get(code);
         if (validator == null) {
@@ -75,8 +86,14 @@ public class ValidatorResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
     public Response post(Validator validator){
-        LOGGER.debug(String.format("Post validator info"));
+        if (!isConfluenceAdmin()) {
+            return Response
+                .status(Response.Status.FORBIDDEN)
+                .entity(resolver.getText("com.mesilat.general.error.permission.rest"))
+                .build();
+        }
 
+        LOGGER.trace(String.format("Create validator %s", validator.getCode()));
         service.create(validator);
         return get(validator.getCode());
     }
@@ -86,22 +103,36 @@ public class ValidatorResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
     public Response put(@PathParam("code") String code, Validator validator){
-        LOGGER.debug(String.format("Update validator info"));
+        if (!isConfluenceAdmin()) {
+            return Response
+                .status(Response.Status.FORBIDDEN)
+                .entity(resolver.getText("com.mesilat.general.error.permission.rest"))
+                .build();
+        }
 
         if (code == null) {
             return Response
                 .status(Response.Status.BAD_REQUEST)
                 .entity("Validator code is not defined")
                 .build();
-        } else {
-            service.update(code, validator);
-            return get(validator.getCode());
         }
+
+        LOGGER.trace(String.format("Update validator %s", code));
+        service.update(code, validator);
+        return get(validator.getCode());
     }
     
     @DELETE
     @Path("/{code}")
     public Response delete(@PathParam("code") String code){
+        if (!isConfluenceAdmin()) {
+            return Response
+                .status(Response.Status.FORBIDDEN)
+                .entity(resolver.getText("com.mesilat.general.error.permission.rest"))
+                .build();
+        }
+
+        LOGGER.trace(String.format("Delete validator %s", code));
         service.delete(code);
         return Response.status(Response.Status.ACCEPTED).build();
     }
@@ -109,8 +140,15 @@ public class ValidatorResource {
     @POST
     @Path("/upload")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response upload(String serialized) throws IOException{
-        LOGGER.debug(String.format("Upload validators"));
+    public Response upload(String serialized) throws IOException {
+        if (!isConfluenceAdmin()) {
+            return Response
+                .status(Response.Status.FORBIDDEN)
+                .entity(resolver.getText("com.mesilat.general.error.permission.rest"))
+                .build();
+        }
+
+        LOGGER.trace("Import validators");
         ArrayNode arr = (ArrayNode)mapper.readTree(serialized);
         arr.forEach(obj -> {
             Validator validator = new Validator(obj.get("code").asText(), obj.has("name")? obj.get("name").asText(): null);
@@ -128,7 +166,13 @@ public class ValidatorResource {
     }
     
     @Inject
-    public ValidatorResource(ValidatorManager service){
+    public ValidatorResource(
+        ValidatorManager service,
+        @ComponentImport PermissionManager permissionManager,
+        @ComponentImport I18nResolver resolver
+    ){
+        super(permissionManager, resolver, null, null);
         this.service = service;
+        this.resolver = resolver;
     }
 }
